chore: remove unused accumulator.rs and new.rs
